Run 1715 NVIDIA master rebase (#12)
* CNT-951 - Update DCGM to 2.1.4, update headers, change default interval

Signed-off-by: Douglas Wightman <dwightman@nvidia.com>

* Update all dockerhub references to nvcr.io

Signed-off-by: Kevin Klues <kklues@nvidia.com>

* Update version refs to v2.1.2 of dcgm-exporter in helm / docs

Everything was tagged a while ago (and container images were released),
but the references in the code weren't updated (nor were any helm charts
released).

Signed-off-by: Kevin Klues <kklues@nvidia.com>

* Bump DCGM Exporter version to 2.2.0 and DCGM versions to 2.1.4

Signed-off-by: Shiva Krishna, Merla <smerla@nvidia.com>

* Add ability to parse old fieldnames and 1.x csv file for backwards compatibility

* Bump aws-kube-ci to 21.02.23

Signed-off-by: Evan Lezar <elezar@nvidia.com>

* DCGM-2116 - Enable -r <HOST>:<PORT> for dcgm-exporter to allow dcgm-exporter to connect to an existing nv-hostengine.

Bump version to 2.3.1

* DCGM-2099. Add bindings to support dcgmGetGpuInstanceHierarchy()

Update the DCGM headers to match DCGM 2.1.5.
Add code to gather the system inventory of GPUs and MIG devices.
When MIG is enabled, add all the MIG devices to the group.

* DCGM-2101. Clearly identify MIG devices in the output for dcgm-exporter

* Simplify some syntax per feedback.

* DCGM-2124. Added instructions for controlling the new arguments in helm.

* DCGM-2103. Add configuration parameters to control what devices are monitored.

Improve the template to only print data that is present.

* Bump the version to 2.4.0

* Set the version to its full name, 2.4.0-rc.2

* DCGM-2155. Add test-main as a makefile target for running go tests.

* Run go mod tidy and go mod vendor

This also removes possibly problematic .travis.yml files from
vendored imports.

Signed-off-by: Evan Lezar <elezar@nvidia.com>

* Issue 187. Fix invalid formatting if logic. Quotes were on the wrong side of the if statements.

Also, add a CLI option (--fake-gpus) to allow testing with fake GPUs.
Last, remove a ':' from saving the GPU id when initializing the MIG
hierarchy.

* Fix using a duplicate MIG entity when publishing metrics.

* DCGM-2234. Ensure that k8s pod information is populated.

The problem was that GPU devices aren't seen when the GPU is in MIG
mode, so we need to create custom identifiers for GPU instances.

* DCGM-2239. Publish GPU-level metrics per GPU instance as well.

Also, no longer offer to report on GPUs and GPU instances at the same
time. Kubernetes libraries see GPU instances as devices, not GPUs when
MIG is enabled. As a result, if metrics aren't reported at the GPU
instance level, the metrics for a GPU are placed in just one pod (even
when two different pods have GPU instances), resulting in an incomplete
picture.

* Needs checking but this bumps the version rebase later

* Update README with repo modification notice

* adding gpu model name to prometheus metric metadata

Signed-off-by: julian3xl <julian3xl@gmail.com>

* Jira validation check

Co-authored-by: Wightman <dwightman@nvidia.com>
Co-authored-by: Pramod Ramarao <pramarao@gmail.com>
Co-authored-by: Kevin Klues <kklues@nvidia.com>
Co-authored-by: Kevin Klues <klueska@gmail.com>
Co-authored-by: Shiva Krishna Merla <smerla@nvidia.com>
Co-authored-by: Evan Lezar <elezar@nvidia.com>
Co-authored-by: Evan Lezar <7723350-elezar@users.noreply.gitlab.com>
Co-authored-by: dbeer <dbeer@nvidia.com>
Co-authored-by: julian3xl <julian3xl@gmail.com>